+++ ProdManager/helpers/form.py
@@ -1,4 +1,7 @@
 
+from flask_wtf import FlaskForm
+from flask import current_app, g
+
 def strip_input(content):
   if content is None:
     return content
@@ -13,3 +16,11 @@ def strip_input(content):
       result.append(line)
 
   return '\n'.join(result)
+
+class CustomForm(FlaskForm):
+  def __init__(self, *args, **kwargs):
+    super().__init__(
+      meta=dict(
+        csrf=current_app.config['WTF_CSRF_ENABLED'] and not g.api
+      ), *args, **kwargs
+    )