+++ tests/ProdManager/helpers/test_helpers_links.py
@@ -1,0 +1,4 @@
+
+from ProdManager.helpers.links import (
+  custom_url_for,
+)